requirements.txt
requests
python-dotenv
discord
asyncio
beautifulsoup4
pytz
seaborn
matplotlib